Fix cooperating with hwcomposer
When hwcomposer is available, framebuffer_close() will be called.
Thus the allocated struct drm_framebuffer is just freed. However,
the gralloc still holds a pointer referring to the freed struct
drm_framebuffer. As a result, the following errors are repeated:

01-11 00:28:47.088  1103  1103 E drm-fb  : Failed to get handle from prime fd: 9
01-11 00:28:48.062  1103  1103 E drm-fb  : Failed to get handle from prime fd: 9

This patch fixes it by setting the pointer to NULL.